pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wudcutt"
version = "0.1.0"
description = "Public-domain woodcut sourcing and apparel-prep workflow"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "Pillow>=10.0.0",
  "numpy>=1.26.0",
  "requests>=2.31.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]

[project.scripts]
wudcutt = "wudcutt.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
